[refactor] Landing: 애니메이션 delay time 상수 파일 생성, delay 변경
