+++ src/app/(landing)/_components/BottomSection/Illustration/Workers.tsx
