+++ src/app/(landing)/_components/MiddleSection/style.ts
