+++ src/app/(landing)/_components/TopSection/Illustration/delayTime.ts
@@ -1,0 +1,1 @@
+export const DELAY_TIME = 1.2;
